pubsub example: interleave calls to publish() with spin_some()
